requirements.txt
numpy>=1.26
Pillow>=10.0
scikit-learn>=1.4
gradio>=4.31
joblib>=1.3
tqdm>=4.66
